requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.8
django-cache-url==3.2.3
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
environs==9.3.4
idna==3.3
marshmallow==3.13.0
PyJWT==2.2.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
